+++ doc/sphinx/changelog.rst
@@ -34,6 +34,7 @@
   * Improve the support and the API for ``LF_ENUM``: |lief-pdb-types-Enum|
   * Improve the support and the API for ``LF_PROCEDURE``: |lief-pdb-types-Function|
   * Improve the support and the API for ``LF_ARRAY``: |lief-pdb-types-Array|
+  * Improve the support and the API for *simple* types: |lief-pdb-types-Simple|
 
 :COFF: